requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langsmith>=0.1.0
boto3>=1.34.0
opensearch-py>=2.4.0
requests-aws4auth>=1.2.0
openai>=1.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
